Reorganizing tests
