+++ .vscode/arduino.json
@@ -1,0 +1,4 @@
+{
+    "board": "arduino:avr:nano",
+    "configuration": "cpu=atmega328old"
+}
