+++ .vscode/c_cpp_properties.json
@@ -1,0 +1,19 @@
+{
+    "configurations": [
+        {
+            "name": "Linux",
+            "includePath": [
+                "/home/skira/Desktop/TG/arduino-1.8.9/tools/**",
+                "/home/skira/Desktop/TG/arduino-1.8.9/hardware/arduino/avr/**"
+            ],
+            "forcedInclude": [
+                "/home/skira/Desktop/TG/arduino-1.8.9/hardware/arduino/avr/cores/arduino/Arduino.h"
+            ],
+            "intelliSenseMode": "gcc-x64",
+            "compilerPath": "/usr/bin/gcc",
+            "cStandard": "c11",
+            "cppStandard": "c++17"
+        }
+    ],
+    "version": 4
+}
