+++ Arduino/Tests_Serial/.vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "python.linting.pylintEnabled": true,
+    "python.linting.enabled": true
+}
