--- Bluetooth/AT Mode/setarBT/setarBT.ino
@@ -1,16 +1,0 @@
-#include <SoftwareSerial.h>
-
-SoftwareSerial BTSerial(9, 10); // RX, TX
-
-void setup() {
-  Serial.begin(9600);
-  Serial.println("Enter AT commands:");
-  BTSerial.begin(38400);
-}
-
-void loop() {
-  if (BTSerial.available())
-    Serial.write(BTSerial.read());
-  if (Serial.available())
-    BTSerial.write(Serial.read());
-}
