--- Bluetooth/GyroComm/gyro/gyro.ino
@@ -1,83 +1,0 @@
-//Programa : Teste MPU-6050
-//Alteracoes e adaptacoes : Arduino e Cia
-//
-//Baseado no programa original de JohnChi
- 
-//Carrega a biblioteca Wire
-#include<Wire.h>
- 
-//GYRO INIT
-
-#include "SoftwareSerial.h"
-SoftwareSerial serial_connection(3,2);//Create a serial connection with TX and RX on these pins
-#define BUFFER_SIZE 64//This will prevent buffer overruns.
-char inData[BUFFER_SIZE];//This is a character buffer where the data sent by the python script will go.
-char inChar=-1;//Initialie the first character as nothing
-int count=0;//This is the number of lines sent in from the python script
-int i=0;//Arduinos are not the most capable chips in the world so I just create the looping variable once
-//GYRO INIT END
-
-
-//Endereco I2C do MPU6050
-const int MPU=0x68;  
-
-//Variaveis para armazenar valores dos sensores
-int AcX,AcY,AcZ,Tmp,GyX,GyY,GyZ;
-
-void setup()
-{
-
-//GYRO MODULE
-  serial_connection.begin(115200);//Initialize communications with the bluetooth module
-  serial_connection.println("Ready!!!");//Send something to just start comms. This will never be seen.
-  Serial.println("Started");//Tell the serial monitor that the sketch has started.
-
-  Serial.begin(115200);
-  Wire.begin();
-  Wire.beginTransmission(MPU);
-  Wire.write(0x6B); 
-   
-  //Inicializa o MPU-6050
-  Wire.write(0); 
-  Wire.endTransmission(true);
-    
-}
-
-void loop()
-{
-  Wire.beginTransmission(MPU);
-  Wire.write(0x3B);  // starting with register 0x3B (ACCEL_XOUT_H)
-  Wire.endTransmission(false);
-  
-  //Solicita os dados do sensor
-  Wire.requestFrom(MPU,14,true);  
-  
-  //Armazena o valor dos sensores nas variaveis correspondentes
-  AcX=Wire.read()<<8|Wire.read(); //0x3B (ACCEL_XOUT_H) & 0x3C (ACCEL_XOUT_L)     
-  AcY=Wire.read()<<8|Wire.read(); //0x3D (ACCEL_YOUT_H) & 0x3E (ACCEL_YOUT_L)
-  AcZ=Wire.read()<<8|Wire.read(); //0x3F (ACCEL_ZOUT_H) & 0x40 (ACCEL_ZOUT_L)
-  Tmp=Wire.read()<<8|Wire.read(); //0x41 (TEMP_OUT_H) & 0x42 (TEMP_OUT_L)
-  GyX=Wire.read()<<8|Wire.read(); //0x43 (GYRO_XOUT_H) & 0x44 (GYRO_XOUT_L)
-  GyY=Wire.read()<<8|Wire.read(); //0x45 (GYRO_YOUT_H) & 0x46 (GYRO_YOUT_L)
-  GyZ=Wire.read()<<8|Wire.read(); //0x47 (GYRO_ZOUT_H) & 0x48 (GYRO_ZOUT_L)
-
-   
-  //Mostra os valores na serial
-  serial_connection.print(AcX, DEC);
-  serial_connection.print(", "); 
-  serial_connection.print(AcY, DEC);
-  serial_connection.print(", "); 
-  serial_connection.print(AcZ, DEC);
-  serial_connection.print("\n"); 
-
-  serial_connection.print(GyX, DEC);
-  serial_connection.print(", "); 
-  serial_connection.print(GyY, DEC);
-  serial_connection.print(", "); 
-  serial_connection.print(GyZ, DEC);
-  serial_connection.print("\n"); 
-  //serial_connection.println(Tmp/340.00+36.53);
-    
-  //Aguarda 300 ms e reinicia o processo
-  delay(200);
-}
